nw, Remove rdpspn16, merge into n64.c since it is only 800 lines or so now
